Lambda Layers

+++ datasources/scripts/_cli.py
@@ -9,7 +9,7 @@ import shutil
 import yaml
 from multiprocessing.pool import ThreadPool
 
-from datasources import Manifest, sources
+from datasources import Manifest, sources, layer_arn
 
 @click.group(short_help="Cognition datasource query")
 def cognition_datasources():
@@ -44,7 +44,6 @@ def search(spatial, start_date, end_date, properties, datasource, debug, output)
 
     manifest = Manifest()
     for source in datasource:
-        manifest.load_source(getattr(sources, source))
         manifest[source].search(geoj, temporal=temporal, properties=properties, limit=10)
 
     if debug:
@@ -64,38 +63,40 @@ def search(spatial, start_date, end_date, properties, datasource, debug, output)
 
     return 0
 
-
 @cognition_datasources.command(name='new')
 @click.option('--name', '-n', type=str)
 def new(name):
     if os.path.exists(name):
         raise ValueError("The directory {} already exists.".format(name))
 
-    shutil.copytree(os.path.join(os.path.dirname(__file__), '..', 'template'), name)
+    shutil.copytree(os.path.join(os.path.dirname(__file__), '..', '..', 'driver'), name)
 
-    with open(os.path.join(os.getcwd(), name, 'template.py'), 'r') as f:
-        contents = f.read()
-        contents = contents.replace('__TEMPLATENAME__', name)
+    fpaths = [
+        os.path.join(os.getcwd(), name, 'template.py'),
+        os.path.join(os.getcwd(), name, 'tests.py'),
+        os.path.join(os.getcwd(), name, 'bin', 'driver-package.sh'),
+        os.path.join(os.getcwd(), name, 'handler.py'),
+        os.path.join(os.getcwd(), name, 'README.md')
+    ]
 
-        with open(os.path.join(os.getcwd(), name, 'template.py'), 'w') as outf:
-            outf.write(contents)
+    for file in fpaths:
+        replace_template_name(file, name)
 
-    with open(os.path.join(os.getcwd(), name, 'tests.py'), 'r') as f:
-        contents = f.read()
-        contents = contents.replace('__TEMPLATENAME__', name)
-
-        with open(os.path.join(os.getcwd(), name, 'tests.py'), 'w') as outf:
-            outf.write(contents)
-
-    os.rename(os.path.join(os.getcwd(), name, 'template.py'), os.path.join(os.getcwd(), name, '{}.py'.format(name)))
-
+    os.rename(fpaths[0], os.path.join(os.path.dirname(fpaths[0]), f'{name}.py'))
 
 @cognition_datasources.command(name='load')
 @click.option('--datasource', '-d', type=str, multiple=True)
-def load(datasource):
+@click.option('--local/--deployed', default=False)
+def load(datasource, local):
+
+    handler = []
+    sls_functions = {}
+
     for source in datasource:
+        print("Loading the {} driver.".format(source))
         source_link = getattr(sources.remote, source)
-        project_path = '/'.join(source_link.split('/')[3:-1])
+        project_path = '/'.join(source_link.split('/')[4:])
+        source_link += '@master'
 
         # Check CI build
         r = requests.get(os.path.join(source_link, 'config.yml'))
@@ -106,45 +107,59 @@ def load(datasource):
             print("WARNING: {} was not loaded because it failed CI".format(source))
             continue
 
-        # Download remote datasource .py file into sources folder
-        source_fname = source + '.py'
-        source_remote_url = os.path.join(source_link, source_fname)
-        r = requests.get(source_remote_url)
-        with open(os.path.join(os.path.dirname(__file__), '..', 'sources', source_fname), 'w+') as outfile:
-            outfile.write(r.text)
+        # Download remote file handler
+        handler_url = os.path.join(source_link, 'handler.py')
+        r = requests.get(handler_url)
+        for line in r.text.splitlines()[2:]:
+            handler.append(line + '\n')
 
-        # Install datasource dependencies
-        fd, path = tempfile.mkstemp()
-        req_remote_url = os.path.join(source_link, 'requirements.txt')
-        try:
-            with os.fdopen(fd, 'w') as tmp:
-                r = requests.get(req_remote_url)
-                tmp.write(r.text)
-        finally:
-            subprocess.call("pip install -r {}".format(path), shell=True)
-            os.remove(path)
+        # Build sls function config
+        sls_functions.update({
+            source: {
+                'handler': 'handler.' + source,
+                'layers': [
+                    layer_arn,
+                    md['layer-arn'],
+                ]
+            }
+        })
 
-        # Check for index
-        idx_remote_url = os.path.join(source_link, 'index.idx')
-        dat_remote_url = os.path.join(source_link, 'index.dat')
+        # Add database layer arn if present
+        if 'db-arn' in md:
+            sls_functions[source]['layers'].append(md['db-arn'])
 
-        idx_r = requests.get(idx_remote_url)
-        dat_r = requests.get(dat_remote_url)
 
-        if idx_r.status_code == 404 or dat_r.status_code == 404:
-            continue
+        if local:
+            # Download driver file to local installation of cognition-datasources
+            driver_url = os.path.join(source_link, f"{source}.py")
+            r = requests.get(driver_url)
+            with open(os.path.join(os.path.dirname(__file__), '..', 'sources', f"{source}.py"), "w+") as driver_file:
+                driver_file.write(r.text)
 
-        static_dir = os.path.join(os.path.dirname(__file__), '..', 'static')
-        if not os.path.exists(static_dir):
-            os.makedirs(static_dir)
+            # Install dependencies
+            fd, path = tempfile.mkstemp()
+            req_url = os.path.join(source_link, "requirements.txt")
+            try:
+                with os.fdopen(fd, 'w') as tmp:
+                    r = requests.get(req_url)
+                    tmp.write(r.text)
+            finally:
+                subprocess.call("pip install -r {}".format(path), shell=True)
+                os.remove(path)
 
-        with open(os.path.join(static_dir, '{}_rtree.idx'.format(source)), 'wb+') as outfile:
-            outfile.write(idx_r.content)
+    # Write handler.py
+    with open(os.path.join(os.path.dirname(__file__), '..', '..', 'handler.py'), 'a+') as outfile:
+        for line in handler:
+            outfile.write(line)
 
-        with open(os.path.join(static_dir, '{}_rtree.dat'.format(source)), 'wb+') as outfile:
-            outfile.write(dat_r.content)
+    # Write serverless.yml
+    with open(os.path.join(os.path.dirname(__file__), '..', '..', 'serverless.yml'), 'r+') as config:
+        contents = yaml.load(config, Loader=yaml.BaseLoader)
+        contents['functions'].update(sls_functions)
 
-        print("Succesfully loaded the {} driver".format(source))
+        with open(os.path.join(os.path.dirname(__file__), '..', '..', 'serverless.yml'), 'w+') as outfile:
+            yaml.dump(contents, outfile)
+
 
 @cognition_datasources.command(name='build-examples')
 def build_examples():
@@ -153,6 +168,7 @@ def build_examples():
     example_rel_path = 'docs/example.json'
 
     def _fetch_examples(data):
+        print("Pulling example for {}.".format(data['name']))
         with open(os.path.join(os.path.dirname(__file__), '..', '..', 'docs', 'examples', '{}.json'.format(data['name'])),
                   'wb+') as examplefile:
             r = requests.get(os.path.join(data['url'], example_rel_path))
@@ -164,12 +180,14 @@ def build_examples():
 @cognition_datasources.command(name='build-docs')
 def build_docs():
     from datasources.sources import remote
+
     remote_assets = {k: v for (k, v) in remote.__dict__.items() if type(v) == str and 'https' in v}
-    docs_rel_path = 'docs/README.md'
+    docs_rel_path = 'README.md'
 
     build_status = []
     with open(os.path.join(os.path.dirname(__file__), '..', '..', 'docs', 'datasource-reference.md'), 'wb+') as docfile:
         for item in remote_assets:
+            print("Pulling docs for {}.".format(item))
             r = requests.get(os.path.join(remote_assets[item], docs_rel_path))
             docfile.write(r.content)
             docfile.write(b"\n---\n")
@@ -178,6 +196,7 @@ def build_docs():
             if 'CircleCI' in lines[0]:
                 build_status.append({'name': item, 'status': lines[0]})
 
+    print("Finishing up.")
     with open(os.path.join(os.path.dirname(__file__), '..', '..', 'docs', 'datasource-status.md'), 'w+') as statusfile:
         statusfile.write("# Driver Status\n")
         statusfile.write("| Driver Name | Status |\n")
@@ -193,3 +212,11 @@ def list():
     print([x.__name__ for x in sources])
 
 
+# Some helper methods used by the CLI
+def replace_template_name(fpath, name):
+    with open(fpath, 'r') as f:
+        contents = f.read()
+        contents = contents.replace('__TEMPLATENAME__', name)
+
+        with open(fpath, 'w') as outf:
+            outf.write(contents)
